Create androidincoming.py
The above API downloads the reports for 2 apps. 
We need to look for bucket id from which we need to download the reports. For eg., in the api, bucket id is the 'pubsite_prod_rev_1048811653366xxxxxxx' and 'pubsite_prod_rev_1048811653366xxxxxxx' and report which we need to download is 'installs_com.justdial.jdlite_201702_overview.csv' and 'installs_com.justdial.search_201702_overview.csv'.


The variable currentMonth stores the value in a yyyymm format. For eg., 201810 stands for October 2018 report.
Googleincoming.log is the file where we can log all the status of the API. If the API fails or passes we log the status in this file.